feat: move route mapping to useCommandBarRoutes
